bugprone-unchecked-string-to-number-conversion Squashed commit of the following:
commit f67a1deab91f451a34c3acb7ebbab568347012aa
Author: Kajol Kulkarni <wu19byjo@i10staff17.informatik.uni-erlangen.de>
Date:   Mon Jun 29 18:59:16 2026 +0200

    removed cerr

commit 1bee2f589d45524f81dbf51daf3feb6f9c0e090d
Author: Kajol Kulkarni <wu19byjo@i10staff17.informatik.uni-erlangen.de>
Date:   Mon Jun 29 18:34:54 2026 +0200

    changed atof to stod and stoul

commit 467381f8cc0add0e6c5ae3ebfa0976ec8cbf1a3a
Author: Kajol Kulkarni <wu19byjo@i10staff17.informatik.uni-erlangen.de>
Date:   Mon Jun 29 18:29:57 2026 +0200

    changed atof to stoul

commit 4b9c86d84b6531605837fef804f0033c2d37987b
Author: Kajol Kulkarni <wu19byjo@i10staff17.informatik.uni-erlangen.de>
Date:   Mon Jun 29 18:29:02 2026 +0200

    changed to atof to stod and stoul

commit b05cc8fb1d6f7cb778862afd1f77625bb7242bf6
Author: Kajol Kulkarni <wu19byjo@i10staff17.informatik.uni-erlangen.de>
Date:   Mon Jun 29 18:24:01 2026 +0200

    changed atof to stoul in HydroDynamicForceOnMultipleBlocks

commit 87e014bcf13af83da9ad94af7077ea68b4cee7bc
Author: Kajol Kulkarni <wu19byjo@i10staff17.informatik.uni-erlangen.de>
Date:   Mon Jun 29 18:17:43 2026 +0200

    changed atof to atof to stoul

commit cae326b04619e21e8a3d94e1ce1bb00e4dfc1b7f
Author: Kajol Kulkarni <wu19byjo@i10staff17.informatik.uni-erlangen.de>
Date:   Mon Jun 29 18:07:26 2026 +0200

    changed atoi to stoi in main

+++ tests/lbm_mesapd_coupling/momentum_exchange_method/DragForceSphereMEM.cpp
@@ -393,8 +393,8 @@ int main( int argc, char **argv )
       if( std::strcmp( argv[i], "--funcTest"  ) == 0 ) { funcTest  = true; continue; }
       if( std::strcmp( argv[i], "--logging"   ) == 0 ) { logging   = true; continue; }
       if( std::strcmp( argv[i], "--MEMVariant") == 0 ) { method    = to_MEMVariant( argv[++i] ); continue; }
-      if( std::strcmp( argv[i], "--tau"       ) == 0 ) { tau       = real_c( std::atof( argv[++i] ) ); continue; }
-      if( std::strcmp( argv[i], "--length"    ) == 0 ) { length    = uint_c( std::atof( argv[++i] ) ); continue; }
+      if( std::strcmp( argv[i], "--tau"       ) == 0 ) { tau       = real_c( std::stod( argv[++i] ) ); continue; }
+      if( std::strcmp( argv[i], "--length"    ) == 0 ) { length    = uint_c( std::stoul( argv[++i] ) ); continue; }
       WALBERLA_ABORT("Unrecognized command line argument found: " << argv[i]);
    }
